Create LoadInspectData.py
Here is some code in python to load the HOF data into a data frame in pandas and inspect quickly inspect it. It provides the first five values and the last five values.  It then displays the count, mean, standard deviation, interquartile ranges, as well as the min and max of each column.